Moved openaddr.ci.collect.iterate_local_processed_files() up to top-level
