+++ openaddr/tests/__init__.py
@@ -21,6 +21,7 @@ import json
 import re
 import pickle
 import sys
+import os
 from os import close, environ, mkdir, remove
 from io import BytesIO
 from csv import DictReader
@@ -41,7 +42,12 @@ from requests import get
 from httmock import response, HTTMock
 import mock
         
-from .. import cache, conform, S3, process_one, package_output
+from .. import (
+    cache, conform, S3, process_one, package_output,
+    iterate_local_processed_files, download_processed_file
+    )
+
+from ..ci.objects import Run
 
 class TestOA (unittest.TestCase):
     
@@ -619,6 +625,66 @@ class TestPackage (unittest.TestCase):
         self.assertEqual(call2[1][1], 'us-ca-carson.txt')
 
         self.assertEqual(call3[0], 'close')
+
+    def test_iterate_local_processed_files(self):
+        state1 = {'processed': 'http://s3.amazonaws.com/openaddresses/123.csv', 'website': 'http://example.com'}
+        state3 = {'processed': 'http://s3.amazonaws.com/openaddresses/789.csv', 'license': 'ODbL'}
+    
+        runs = [
+            Run(123, 'sources/123.json', 'abc', b'', None,
+                state1, None, None, None, None, None, None, None),
+            Run(456, 'sources/456.json', 'def', b'', None,
+                {'processed': None}, None, None, None, None, None, None, None),
+            Run(789, 'sources/7/9.json', 'ghi', b'', None,
+                state3, None, None, None, None, None, None, None),
+            ]
+        
+        with mock.patch('openaddr.download_processed_file') as download_processed_file:
+            download_processed_file.return_value = 'nonexistent file'
+            local_processed_files = iterate_local_processed_files(runs)
+            
+            self.assertEqual(next(local_processed_files), ('123', 'nonexistent file', state1))
+            self.assertEqual(next(local_processed_files), ('7/9', 'nonexistent file', state3))
+
+    def response_content(self, url, request):
+        '''
+        '''
+        MHP = request.method, url.hostname, url.path
+        
+        if MHP == ('GET', 's3.amazonaws.com', '/openaddresses/us-oh-clinton.csv'):
+            return response(200, b'...', headers={'Last-Modified': 'Wed, 30 Apr 2014 17:42:10 GMT'})
+        
+        if MHP == ('GET', 'data.openaddresses.io.s3.amazonaws.com', '/runs/11170/ca-ab-strathcona-county.zip'):
+            return response(200, b'...', headers={'Last-Modified': 'Tue, 18 Aug 2015 07:10:32 GMT'})
+        
+        if MHP == ('GET', 'data.openaddresses.io.s3.amazonaws.com', '/runs/13616/fr/vaucluse.zip'):
+            return response(200, b'...', headers={'Last-Modified': 'Wed, 19 Aug 2015 10:35:44 GMT'})
+
+        raise ValueError(url.geturl())
+        
+    def test_download_processed_file_csv(self):
+        with HTTMock(self.response_content):
+            filename = download_processed_file('http://s3.amazonaws.com/openaddresses/us-oh-clinton.csv')
+
+        self.assertEqual(splitext(filename)[1], '.csv')
+        self.assertEqual(os.stat(filename).st_mtime, 1398879730)
+        remove(filename)
+
+    def test_download_processed_file_zip(self):
+        with HTTMock(self.response_content):
+            filename = download_processed_file('http://data.openaddresses.io.s3.amazonaws.com/runs/11170/ca-ab-strathcona-county.zip')
+
+        self.assertEqual(splitext(filename)[1], '.zip')
+        self.assertEqual(os.stat(filename).st_mtime, 1439881832)
+        remove(filename)
+
+    def test_download_processed_file_nested_zip(self):
+        with HTTMock(self.response_content):
+            filename = download_processed_file('http://data.openaddresses.io.s3.amazonaws.com/runs/13616/fr/vaucluse.zip')
+
+        self.assertEqual(splitext(filename)[1], '.zip')
+        self.assertEqual(os.stat(filename).st_mtime, 1439980544)
+        remove(filename)
 
 @contextmanager
 def locked_open(filename):
